+++ tests/Console/Application/ParallelProcessesApplication/ConstructorTest.php
@@ -20,5 +20,6 @@ final class ConstructorTest extends TestCase
         static::assertCount(0, $application->getProcesses());
         static::assertInstanceOf(DefaultTheme::class, $application->getTheme());
         static::assertSame(10000, $application->getRefreshInterval());
+        static::assertNull($application->getMaximumParallelProcesses());
     }
 }